Splitted SelectableTable content into multiple components for clearer understanding
